requirements.txt
fastapi
uvicorn
pymongo
sqlalchemy
python-dotenv
pydantic
langchain
langchain-core
langchain-openai
openai
uuid
typing_extensions
mysql-connector-python
dramatiq
psutil
prometheus-client